rename projects DSL declaration in yaml for code coherence
