+++ projects.go
@@ -2,74 +2,90 @@ package main
 
 import (
 	"fmt"
-	"net/url"
+	"log"
+	"os"
+	"path"
+	"strings"
 
-	"github.com/forj-oss/forjj-modules/trace"
 	"github.com/forj-oss/goforjj"
 )
 
-type ProjectInfo struct {
-	ForjCommonStruct
-	dslRepo         string
-	dslPath         string
-	isDslDefault    bool
-	IsProDeployment bool `yaml:"production-deployment,omitempty"`
+type Projects struct {
+	DslRepo    string `yaml:"dsl-repo,omitempty"`
+	DslPath    string `yaml:"dsl-path,omitempty"`
+	DslDefault bool   `yaml:"dsl-default,omitempty"`
+	infra_name string
+	List       map[string]Project
 }
 
-func (pi *ProjectInfo) set_project_info(forj ForjCommonStruct) {
-	pi.ForjCommonStruct = forj
+func NewProjects(InstanceName, repo, Dslpath string, dslDefault bool) *Projects {
+	p := new(Projects)
+	p.DslPath = Dslpath
+	p.DslRepo = repo
+
+	p.DslDefault = dslDefault
+	p.List = make(map[string]Project)
+	return p
 }
 
-func (pi *ProjectInfo) setDslInfo(seedJob SeedJobStruct) {
-	if seedJob.Path == "" && seedJob.Repo == "" {
-		pi.isDslDefault = true
-		pi.dslRepo = seedJob.DefaultRepo
-		pi.dslPath = seedJob.DefaultPath
-	} else {
-		pi.dslRepo = seedJob.Repo
-		pi.dslPath = seedJob.Path
-	}
+func (p *Projects) AddGithub(name string, d *GithubStruct) bool {
+	data := new(GithubStruct)
+	data.SetFrom(d)
+	p.List[name] = Project{Name: name, SourceType: "github", Github: *data, all: p}
+	return true
 }
 
-func (pi *ProjectInfo) setIsProDeploy(isProDeployment bool) {
-	pi.IsProDeployment = isProDeployment
+func (p *Projects) AddGit(name string, d *GitStruct) bool {
+	data := new(GitStruct)
+	data.SetFrom(d)
+	p.List[name] = Project{Name: name, SourceType: "git", Git: *data, all: p}
+	return true
 }
 
-func (pi *ProjectInfo) set_projects_to(projects map[string]ProjectsInstanceStruct, r *JenkinsPlugin,
-	ret *goforjj.PluginData, status *bool, InfraName string) (_ error) {
-	if pi.ForjjInfraUpstream == "" {
-		ret.StatusAdd("Unable to add a new project without a remote GIT repository. Jenkins JobDSL requirement. " +
-			"To enable this feature, add a remote GIT to your infra --infra-upstream or define the JobDSL Repository to clone.")
-		IsUpdated(status)
+// Generates Jobs-dsl files in the given checked-out GIT repository.
+func (p *Projects) Generates(jp *JenkinsPlugin, instance_name string, ret *goforjj.PluginData, status *bool) (_ error) {
+	if !p.DslDefault {
+		log.Printf(ret.StatusAdd("Deploy: JobDSL groovy files ignored. To generate them, unset seed-job-repo and seed-job-path."))
 		return
 	}
 
-	if v, err := url.Parse(pi.dslRepo); err != nil {
-		ret.Errorf("JobDSL: Infra remote URL issue. Check `seed-job-repo` parameter given by Forjj. %s", err)
+	template_dir := jp.template_dir
+	repo_path := jp.deployPath
+
+	if f, err := os.Stat(repo_path); err != nil {
 		return err
 	} else {
-		if v.Scheme == "" {
-			err = fmt.Errorf("JobDSL: Invalid Remote repository Url '%s'. Check `seed-job-repo` parameter given by Forjj. A scheme must exist", pi.dslRepo)
-			ret.Errorf("%s", err)
+		if !f.IsDir() {
+			return fmt.Errorf(ret.Errorf("Repo cloned path '%s' is not a directory.", repo_path))
+		}
+	}
+
+	jobs_dsl_path := path.Join(repo_path, p.DslPath)
+	if f, err := os.Stat(jobs_dsl_path); err != nil {
+		if err := os.MkdirAll(jobs_dsl_path, 0755); err != nil {
 			return err
 		}
+	} else {
+		if !f.IsDir() {
+			return fmt.Errorf(ret.Errorf("path '%s' is not a directory.", repo_path))
+		}
 	}
-	// Initialize JobDSL structure
-	r.yaml.Projects = NewProjects(pi.ForjjInstanceName, pi.dslRepo, pi.dslPath, pi.isDslDefault)
 
-	// Retrieve list of Repository (projects) to manage
-	for name, prj := range projects {
-		if prj.RepoDeployHosted != "true" {
-			gotrace.Trace("Project %s ignored, because not deploying in production an '%s' repo role.", name, prj.RepoRole)
-			continue
-		}
-		switch prj.RemoteType {
-		case "github":
-			r.yaml.Projects.AddGithub(name, &prj.GithubStruct)
-		case "git":
-			r.yaml.Projects.AddGit(name, &prj.GitStruct)
+	tmpl := new(TmplSource)
+	tmpl.Template = "jobs-dsl/job-dsl.groovy"
+	tmpl.Chmod = 0644
+
+	for name, prj := range p.List {
+		name = strings.Replace(name, "-", "_", -1)
+		if u, err := tmpl.Generate(prj.Model(jp), template_dir, jobs_dsl_path, name+".groovy"); err != nil {
+			log.Printf("Deploy: Unable to generate '%s'. %s",
+				path.Join(jobs_dsl_path, name+".groovy"), ret.Errorf("%s", err))
+			return err
+		} else if u {
+			IsUpdated(status)
+			ret.AddFile(goforjj.FilesDeploy, path.Join(instance_name, p.DslPath, name+".groovy"))
+			log.Printf(ret.StatusAdd("Deploy: Project '%s' (%s) generated", name, path.Join(p.DslPath, name+".groovy")))
 		}
 	}
-	IsUpdated(status)
-	return
+	return nil
 }
